5 days forecast

+++ app/src/main/java/com/example/weather/model/Main.java
@@ -2,7 +2,7 @@ package com.example.weather.model;
 
 public class Main {
     private double temp;
-    private int pressure ;
+    private double pressure ;
     private int humidity;
     private double temp_min;
     private double temp_max;
@@ -18,11 +18,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
